# internal/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BevShop API",
    "version": "1.0.0",
    "description": "Single-owner beverage shop API. Use auth/setup once, or auth/login, then copy the returned token into Authorize (without the Bearer prefix). Request IDs must reference existing records. Dates use RFC3339 timestamps with an explicit timezone. This document describes the currently implemented routes."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/suppliers": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "summary": "List suppliers",
        "operationId": "supplier_list",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/supplier_Supplier"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Create supplier",
        "operationId": "supplier_create",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/supplier_Supplier"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate record, record in use, or insufficient stock.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/supplier_CreateRequest"
              },
              "example": {
                "name": "Main Supplier",
                "phone": "0780000000",
                "email": "owner@example.com",
                "address": "Kigali",
                "notes": ""
              }
            }
          }
        }
      }
    },
    "/api/v1/suppliers/{id}": {
      "put": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Update supplier",
        "operationId": "supplier_update",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/supplier_Supplier"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate record, record in use, or insufficient stock.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/supplier_UpdateRequest"
              },
              "example": {
                "name": "Main Supplier",
                "phone": "0780000000",
                "email": "owner@example.com",
                "address": "Kigali",
                "notes": ""
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Delete supplier",
        "operationId": "supplier_delete",
        "responses": {
          "204": {
            "description": "Deleted; no response body."
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate record, record in use, or insufficient stock.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the token from setup or login."
      }
    },
    "schemas": {
...
      "supplier_Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "name": {
            "type": "string",
            "example": "Main Supplier"
          },
          "phone": {
            "type": "string",
            "example": "0780000000"
          },
          "email": {
            "type": "string",
            "example": "owner@example.com"
          },
          "address": {
            "type": "string",
            "example": "Kigali"
          },
          "notes": {
            "type": "string",
            "example": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          }
        }
      },
      "supplier_CreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main Supplier",
            "minLength": 1
          },
          "phone": {
            "type": "string",
            "example": "0780000000"
          },
          "email": {
            "type": "string",
            "example": "owner@example.com"
          },
          "address": {
            "type": "string",
            "example": "Kigali"
          },
          "notes": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "name"
        ]
      },
      "supplier_UpdateRequest": {
        "$ref": "#/components/schemas/supplier_CreateRequest"
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid request body"
          }
        }
      },
...
    }
  }
}